feat: modularize display and style apply/cleanup/dry-run
