+++ src/lib/ux/constants.ts
@@ -2,6 +2,7 @@ import chalk from 'chalk';
 
 // Letta brand color
 export const LETTA_PURPLE = '#7C7CFF';
+export const purple = chalk.hex(LETTA_PURPLE);
 
 // ASCII art banner
 export const BANNER = `
