+++ src/lib/ux/display/index.ts
@@ -1,0 +1,31 @@
+// Resource list displays
+export {
+  displayAgents, AgentData,
+  displayBlocks, BlockData,
+  displayTools, ToolData,
+  displayFolders, FolderData,
+  displayMcpServers, McpServerData,
+  displayFiles, FileData,
+} from './resources';
+
+// Detail/describe displays
+export {
+  displayAgentDetails, AgentDetailsData,
+  displayBlockDetails, BlockDetailsData,
+  displayToolDetails, ToolDetailsData,
+  displayFolderDetails, FolderDetailsData,
+  displayFileDetails, FileDetailsData,
+  displayMcpServerDetails, McpServerDetailsData,
+} from './details';
+
+// Message history display
+export { displayMessages, MessageDisplayData } from './messages';
+
+// Apply summary display
+export { displayApplySummary, ApplySummaryData } from './apply';
+
+// Cleanup display
+export { displayOrphanedResources, displayCleanupNote, OrphanedItem } from './cleanup';
+
+// Dry-run display
+export { displayDryRunSeparator, displayDryRunSummary, displayDryRunAction, DryRunSummaryData } from './dry-run';
